stub out simple starting points for columnizers
Some of the base architecture in place with a simple columnizer that creates a
fixed number of columns without regard for actual height, etc, simply splitting
child elements into each of the columns.

This obviously isn't what the columnizer really needs to do (and it likely does
not even handle text nodes as direct children of the selector correctly), but
gives us a starting point for the design of settings validation, columnization,
etc.