trying to fix things that failed in travis
Build 1 failed because of invalid reference to jQuery version. Since that was
in the node package.json file I don't think we really need it since we don't
actually have a dependency within node itself on jQuery. Rather, we have it at
runtime only and we are packaging it along with our code, so shouldn't need it
in the node env.